Eliminate some narrowing conversion warnings

+++ src/lib/libcmd/uname.c
@@ -127,7 +127,7 @@ static bool output(bool sep, uint32_t flags, uint32_t flag, const char *value, c
 }
 
 int b_uname(int argc, char **argv, Shbltin_t *context) {
-    long flags = 0;
+    uint32_t flags = 0;
     bool sep = false;
     int n;
     char *s;
@@ -206,7 +206,7 @@ int b_uname(int argc, char **argv, Shbltin_t *context) {
         __builtin_unreachable();
     }
     if (sethost) {
-        if (sethostname(sethost, strlen(sethost) + 1)) {
+        if (sethostname(sethost, (int)strlen(sethost) + 1)) {
             error(ERROR_system(1), "%s: cannot set host name", sethost);
             __builtin_unreachable();
         }
